refacto(server): move time_elapsed in utils file
