+++ server/include/utils.h
@@ -8,6 +8,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <sys/time.h>
 
 /**
  * @brief Parse a positive int argument.
@@ -17,3 +18,11 @@
  * CODE_ERROR_INVALID_NUMBER otherwise.
  */
 int parse_positive_int_arg(char *arg);
+
+/**
+ * @brief Calcul time elapsed whereas time parameter.
+ *
+ * @param time Time of depart.
+ * @return double Time.
+ */
+double time_elapsed(struct timeval *time);
